R6 Class for OSV Querybatch Endpoint

An R6 class to provide a lower-level interface to the querybatch endpoint of the OSV API. Batches are enforced to only process by commit hash, purl, or name+ecosystem. This avoids some confusion as to which is taken preferentially and simplifies query creation.

Value

An R6 object to operate with OSV querybatch endpoint.

Details

Pageination is not implemented yet, waiting for httr2 updates to add, then will handle automatically. For now, a warning will be provided and the first set returned. The response object can be used for subsequent queries if the user wants to handle the pagination themselves for the time being.

Super class

rosv::RosvQuery1 -> RosvQueryBatch

Methods

Public methods

RosvQueryBatch$new()

RosvQueryBatch$run()

RosvQueryBatch$parse()

RosvQueryBatch$clone()

Inherited methods

rosv::RosvQuery1$print()

Method new()

Set the core request details for subsequent use when called in run() method.

Usage

RosvQueryBatch$new(
  commit = NULL,
  version = NULL,
  name = NULL,
  ecosystem = NULL,
  purl = NULL,
  page_token = NULL
)

Arguments

commit

Commit hash to query against (do not use when version set).

version

Version of package.

name

Name of package.

ecosystem

Ecosystem package lives within (must be set if using name).

purl

URL for package (do not use if name or ecosystem set).

page_token

When large number of results, next response to complete set requires a page_token.

Method run()

Perform the request and return response for OSV API call.

Usage

RosvQueryBatch$run()

Method parse()

Parse the contents returned into a tidier format.

Usage

RosvQueryBatch$parse()

Details

When no result is found, any empty list is returned by the API, which during parsing will be dropped as the list is flattened. However, the index of the list is still accessible and the dropped items can easily be identified from the results column.

Method clone()

The objects of this class are cloneable with this method.

Usage

RosvQueryBatch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

pkgs <- c('jinja2', 'dask')
ecosystem <- rep('PyPI', length(pkgs))
batchquery <- RosvQueryBatch$new(name = pkgs, ecosystem = ecosystem)
batchquery
#> Request made to: NA 
#> Response status of: NA 
#> Content length is: NA