R6 Class for OSV Vulns Endpoint

An R6 class to provide a lower-level interface to the vulnerability endpoint of the OSV API.

Value

An R6 object to operate with OSV vulns endpoint.

Super class

rosv::RosvQuery1 -> RosvVulns

Methods

Public methods

RosvVulns$new()

RosvVulns$run()

RosvVulns$print()

RosvVulns$clone()

Inherited methods

rosv::RosvQuery1$parse()

Method new()

Set the core request details for subsequent use when called in run() method.

Usage

RosvVulns$new(vuln_ids)

Arguments

vuln_ids

Character vector of vulnerability IDs.

Method run()

Perform the request and return response for OSV API call.

Usage

RosvVulns$run()

Method print()

Print basic details of query object to screen.

Usage

RosvVulns$print(...)

Arguments

...

Reserved for possible future use.

Method clone()

The objects of this class are cloneable with this method.

Usage

RosvVulns$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

vulns <- RosvQueryBatch$new(c('RSEC-2023-6', 'GHSA-jq35-85cj-fj4p'))
vulns
#> Request made to: NA 
#> Response status of: NA 
#> Content length is: NA