Check input against possible ecosystems available

Ensures that inputs for ecosystem are valid based upon what is available in OSV database.

Usage

check_ecosystem(ecosystem, suppressMessages = TRUE)

Arguments

ecosystem

Character value for ecosystem(s) to check.

suppressMessages

Boolean value whether or not to suppress any messages.

Value

A character vector of the same input if all are valid ecosystem names.

Details

Will attempt to grab latest file and cache for the session. If cannot access the online version, will use a local copy that is shipped with the package.

See also

Examples

# Passes
rosv:::check_ecosystem(c('PyPI', 'CRAN'))
#> [1] "PyPI" "CRAN"

# Fails
try(rosv:::check_ecosystem(c('notvalid', 'pypi')))
#> Error in purrr::map_chr(ecosystem, function(x) match.arg(x, ecosystems$ecosystem,  : 
#>   ℹ In index: 1.
#> Caused by error in `match.arg()`:
#> ! 'arg' should be one of “AlmaLinux”, “AlmaLinux:8”, “AlmaLinux:9”, “Alpine”, “Alpine:v3.10”, “Alpine:v3.11”, “Alpine:v3.12”, “Alpine:v3.13”, “Alpine:v3.14”, “Alpine:v3.15”, “Alpine:v3.16”, “Alpine:v3.17”, “Alpine:v3.18”, “Alpine:v3.2”, “Alpine:v3.3”, “Alpine:v3.4”, “Alpine:v3.5”, “Alpine:v3.6”, “Alpine:v3.7”, “Alpine:v3.8”, “Alpine:v3.9”, “Android”, “Bitnami”, “CRAN”, “Debian”, “Debian:10”, “Debian:11”, “Debian:12”, “Debian:3.0”, “Debian:3.1”, “Debian:4.0”, “Debian:5.0”, “Debian:6.0”, “Debian:7”, “Debian:8”, “Debian:9”, “GSD”, “GitHub”, “Actions”, “Go”, “Hackage”, “Hex”, “Linux”, “Maven”, “NuGet”, “OSS-Fuzz”, “Packagist”, “Pub”, “PyPI”, “Rocky”, “Linux:8”, “Linux:9”, “RubyGems”, “SwiftURL”, “UVI”, “crates.io”, “npm”