Create blacklist commands for Posit Package Manager

Use OSV data accessed via osv_query to create blacklist (i.e. blocklist) commands for the Posit Package Manager product.

Usage

create_ppm_blacklist(rosv_query, flags = NULL)

Arguments

rosv_query

A table of vulnerabilities (created via osv_query()).

flags

Global flag to append to commands.

Value

Character vector containing blacklist commands.

Details

Although OSV has many databases for open source software, this function is only relevant for CRAN/Bioconductor and PyPI. To ensure the blacklist is applied to the appropriate target, it is encouraged to specify the name of the source used in your configuration as an additional flag parameter (see examples). Only one ecosystem can be used at a time to ensure there is not a mix of packages across ecosystems applied to incompatible sources.

Examples

if (FALSE) { # interactive()

# Blacklist all CRAN package versions with a listed vulnerability
cran_vul <- osv_query(ecosystem = 'CRAN', all_affected = FALSE)
cmd_blist <- create_ppm_blacklist(cran_vul, flags = '--source=cran')
}