Fetch all available ecosystems

Internal function used to fetch the available ecosystems in the OSV API.

Usage

fetch_ecosystems(offline = FALSE, refresh = FALSE)

Arguments

offline

Boolean, determine if use list bundled with package.

refresh

Boolean, force refresh of cache when using online list.

Value

A data.frame containing all the ecosystem names available in the OSV database.

Details

The refresh parameter can be used to force the data to be pulled again even if one is available in the cached location. Since a fresh pull is performed for each R session, it is unlikely that this parameter is required and is primarily reserved for future use if functionality necessitates.

See also

Examples

rosv:::fetch_ecosystems(offline = TRUE)
#>       ecosystem
#> 1     AlmaLinux
#> 2   AlmaLinux:8
#> 3   AlmaLinux:9
#> 4        Alpine
#> 5  Alpine:v3.10
#> 6  Alpine:v3.11
#> 7  Alpine:v3.12
#> 8  Alpine:v3.13
#> 9  Alpine:v3.14
#> 10 Alpine:v3.15
#> 11 Alpine:v3.16
#> 12 Alpine:v3.17
#> 13 Alpine:v3.18
#> 14  Alpine:v3.2
#> 15  Alpine:v3.3
#> 16  Alpine:v3.4
#> 17  Alpine:v3.5
#> 18  Alpine:v3.6
#> 19  Alpine:v3.7
#> 20  Alpine:v3.8
#> 21  Alpine:v3.9
#> 22      Android
#> 23      Bitnami
#> 24         CRAN
#> 25       Debian
#> 26    Debian:10
#> 27    Debian:11
#> 28    Debian:12
#> 29   Debian:3.0
#> 30   Debian:3.1
#> 31   Debian:4.0
#> 32   Debian:5.0
#> 33   Debian:6.0
#> 34     Debian:7
#> 35     Debian:8
#> 36     Debian:9
#> 37          GSD
#> 38       GitHub
#> 39      Actions
#> 40           Go
#> 41      Hackage
#> 42          Hex
#> 43        Linux
#> 44        Maven
#> 45        NuGet
#> 46     OSS-Fuzz
#> 47    Packagist
#> 48          Pub
#> 49         PyPI
#> 50        Rocky
#> 51        Linux
#> 52        Rocky
#> 53      Linux:8
#> 54        Rocky
#> 55      Linux:9
#> 56     RubyGems
#> 57          UVI
#> 58    crates.io
#> 59          npm